+++ crates/metadata-store/src/network/mod.rs
@@ -1,0 +1,20 @@
+// Copyright (c) 2023 - 2024 Restate Software, Inc., Restate GmbH.
+// All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+mod connection_manager;
+mod grpc_svc;
+mod handler;
+mod networking;
+
+pub use connection_manager::ConnectionManager;
+pub use grpc_svc::raft_metadata_store_svc_server::RaftMetadataStoreSvcServer;
+pub use grpc_svc::FILE_DESCRIPTOR_SET;
+pub use handler::RaftMetadataStoreHandler;
+pub use networking::Networking;
